feat 3.27: added emails to accept/decline

+++ BookingApplication/src/main/java/com/BookingApp/service/AppUserService.java
@@ -6,6 +6,8 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -25,6 +27,8 @@ public class AppUserService {
 	private UserRepository userRepository;
 	@Autowired
 	private RequestDeleteAccRepository requestDeleteAccRepository;
+	@Autowired
+	private JavaMailSender javaMailSender;
 	
 	
 	@GetMapping(path="/getUsers")
@@ -68,16 +72,74 @@ public class AppUserService {
 	@PostMapping(path="/acceptRequest")
 	public ResponseEntity<List<RequestDeleteAcc>> acceptRequest(@RequestBody RequestDeleteAcc request)
 	{	
-		userRepository.deleteById(request.appUserId);
-		requestDeleteAccRepository.delete(request);
-		return new ResponseEntity<List<RequestDeleteAcc>>(requestDeleteAccRepository.findAll(),HttpStatus.OK);
+		String body = "Hello,\nYour request for account deletion has been accepted.\n" 
+				  + "\nIf you have any trouble, write to our support : isa.projekat.tester@gmail.com";
+		Optional<AppUser> oldUser = userRepository.findById(request.appUserId);
+		AppUser appUser;
+		if(oldUser.isPresent()) {
+			appUser = oldUser.get();
+			try 
+			{
+				Thread t = new Thread() {
+					public void run()
+					{
+						sendEmail(appUser.email,body);		
+					}
+				};
+				t.start();
+				userRepository.deleteById(request.appUserId);
+				requestDeleteAccRepository.delete(request);
+				return new ResponseEntity<List<RequestDeleteAcc>>(requestDeleteAccRepository.findAll(),HttpStatus.OK);
+			} 
+			catch (Exception e) 
+			{
+				return null;
+			}
+		}			
+		else
+			return null;
 	}
 	
 	
 	@PostMapping(path="/declineRequest")
 	public ResponseEntity<List<RequestDeleteAcc>> declineRequest(@RequestBody RequestDeleteAcc request)
 	{	
-		requestDeleteAccRepository.delete(request);
-		return new ResponseEntity<List<RequestDeleteAcc>>(requestDeleteAccRepository.findAll(),HttpStatus.OK);
+		String body = "Hello,\nYour request for account deletion has been declined.\n" 
+				  + "\nIf you have any trouble, write to our support : isa.projekat.tester@gmail.com";
+		Optional<AppUser> oldUser = userRepository.findById(request.appUserId);
+		AppUser appUser;
+		if(oldUser.isPresent()) {
+			appUser = oldUser.get();
+			try 
+			{
+				Thread t = new Thread() {
+					public void run()
+					{
+						sendEmail(appUser.email,body);		
+					}
+				};
+				t.start();
+			requestDeleteAccRepository.delete(request);
+			return new ResponseEntity<List<RequestDeleteAcc>>(requestDeleteAccRepository.findAll(),HttpStatus.OK);
+			} 
+			catch (Exception e) 
+			{
+				return null;
+			}
+		}			
+		else
+			return null;
+	}
+	
+	
+	public void sendEmail(String to, String body)
+	{
+		String title = "Request for account deletion";
+		SimpleMailMessage msg = new SimpleMailMessage();
+		msg.setTo(to);
+		msg.setSubject(title);
+		msg.setText(body);
+		javaMailSender.send(msg);
+		System.out.println("Email sent...");
 	}
 }
